examples/twilio/call/python/requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
bidict==0.23.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
elevenlabs==1.1.2
exceptiongroup==1.2.1
executing==2.0.1
Flask==3.0.3
Flask-SocketIO==5.3.6
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipython==8.23.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
multidict==6.0.5
parso==0.8.4
prompt-toolkit==3.0.43
pure-eval==0.2.2
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
PyJWT==2.8.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
traitlets==5.14.3
twilio==9.0.5
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.2
wsproto==1.2.0
yarl==1.9.4
